refactor: use centralized logging in update scripts
